Passing the bitmap saving to viewModel
