+++ app/src/main/java/com/dev/agalperin/presentation/HomeScreenEffect.kt
@@ -5,6 +5,6 @@ import com.dev.agalperin.utils.ErrorType
 
 sealed class HomeScreenEffect {
     data class ShowError(val error: ErrorType): HomeScreenEffect()
-    data class ShowImageSavedSnackbar(val uri: Uri) : HomeScreenEffect()
+    data class ShowImageSavedSnackbar(val uri: Uri?) : HomeScreenEffect()
 }
